+++ wikibot/src/main/java/com/github/wikibot/dumps/StAXDumpReader.java
@@ -133,7 +133,7 @@ public final class StAXDumpReader implements Iterable<XMLRevision>, AutoCloseabl
 	private static class PageInfo {
 		String title;
 		int ns;
-		int id;
+		long id;
 		boolean isRedirect;
 		
 		void clear() {
@@ -167,7 +167,7 @@ public final class StAXDumpReader implements Iterable<XMLRevision>, AutoCloseabl
 					page.ns = Integer.parseInt(sb.toString());
 					break;
 				case "id":
-					page.id = Integer.parseInt(sb.toString());
+					page.id = Long.parseLong(sb.toString());
 					break;
 				case "redirect":
 					page.isRedirect = true;
@@ -189,11 +189,11 @@ public final class StAXDumpReader implements Iterable<XMLRevision>, AutoCloseabl
 				case "id":
 					// ignore if already set, e.g. contributor's id
 					if (revision.revid == 0) {
-						revision.revid = Integer.parseInt(sb.toString());
+						revision.revid = Long.parseLong(sb.toString());
 					}
 					break;
 				case "parentid":
-					revision.parentid = Integer.parseInt(sb.toString());
+					revision.parentid = Long.parseLong(sb.toString());
 					break;
 				case "timestamp":
 					revision.timestamp = sb.toString();
